pyproject.toml
[tool.poetry]
name = "homework"
version = "0.1.0"
description = ""
authors = ["Твое Имя <email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"

[tool.poetry.group.lint.dependencies]
flake8 = "^6.1.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.4.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
)/
'''

[tool.isort]
line_length = 119
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
force_grid_wrap = 0
combine_as_imports = true
lines_between_types = 1
lines_after_imports = 2

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'